Move Options struct to new module
