+++ src/hyperfine/mod.rs
@@ -5,6 +5,7 @@ pub mod error;
 pub mod export;
 pub mod format;
 pub mod internal;
+pub mod options;
 pub mod outlier_detection;
 pub mod parameter_range;
 pub mod shell;
