+++ src/hyperfine/types.rs
@@ -6,7 +6,7 @@ use std::fmt;
 use rust_decimal::prelude::ToPrimitive;
 use rust_decimal::Decimal;
 
-use crate::hyperfine::units::{Second, Unit};
+use crate::hyperfine::units::Second;
 
 #[cfg(not(windows))]
 pub const DEFAULT_SHELL: &str = "sh";
@@ -67,106 +67,6 @@ impl<'a> ToString for ParameterValue {
         match self {
             ParameterValue::Text(ref value) => value.clone(),
             ParameterValue::Numeric(value) => value.to_string(),
-        }
-    }
-}
-
-/// Action to take when an executed command fails.
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum CmdFailureAction {
-    /// Exit with an error message
-    RaiseError,
-
-    /// Simply ignore the non-zero exit code
-    Ignore,
-}
-
-/// Output style type option
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum OutputStyleOption {
-    /// Do not output with colors or any special formatting
-    Basic,
-
-    /// Output with full color and formatting
-    Full,
-
-    /// Keep elements such as progress bar, but use no coloring
-    NoColor,
-
-    /// Keep coloring, but use no progress bar
-    Color,
-
-    /// Disable all the output
-    Disabled,
-}
-
-/// Number of runs for a benchmark
-pub struct Runs {
-    /// Minimum number of benchmark runs
-    pub min: u64,
-
-    /// Maximum number of benchmark runs
-    pub max: Option<u64>,
-}
-
-impl Default for Runs {
-    fn default() -> Runs {
-        Runs { min: 10, max: None }
-    }
-}
-
-/// A set of options for hyperfine
-pub struct HyperfineOptions {
-    /// Number of warmup runs
-    pub warmup_count: u64,
-
-    /// Number of benchmark runs
-    pub runs: Runs,
-
-    /// Minimum benchmarking time
-    pub min_time_sec: Second,
-
-    /// Whether or not to ignore non-zero exit codes
-    pub failure_action: CmdFailureAction,
-
-    /// Command to run before each timing run
-    pub preparation_command: Option<Vec<String>>,
-
-    /// Command to run after each benchmark
-    pub cleanup_command: Option<String>,
-
-    /// What color mode to use for output
-    pub output_style: OutputStyleOption,
-
-    /// The shell to use for executing commands.
-    pub shell: String,
-
-    /// Forward benchmark's stdout to hyperfine's stdout
-    pub show_output: bool,
-
-    /// Which time unit to use for CLI & Markdown output
-    pub time_unit: Option<Unit>,
-
-    /// A list of custom command names that, if defined,
-    /// will be used instead of the command itself in
-    /// benchmark outputs.
-    pub names: Option<Vec<String>>,
-}
-
-impl Default for HyperfineOptions {
-    fn default() -> HyperfineOptions {
-        HyperfineOptions {
-            names: None,
-            warmup_count: 0,
-            runs: Runs::default(),
-            min_time_sec: 3.0,
-            failure_action: CmdFailureAction::RaiseError,
-            preparation_command: None,
-            cleanup_command: None,
-            output_style: OutputStyleOption::Full,
-            shell: DEFAULT_SHELL.to_string(),
-            show_output: false,
-            time_unit: None,
         }
     }
 }
